requirements.txt
# Django and core
Django==6.0.1
psycopg2-binary==2.9.9  # For PostgreSQL
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0  # For production

# Database and caching
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Forms and templates
crispy-bootstrap5==2023.10
django-crispy-forms==2.1
django-widget-tweaks==1.5.0

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
markdown==3.5.1
drf-yasg==1.21.7  # API documentation

# Image processing
#Pillow==10.1.0
django-imagekit==5.0.0

# File storage
boto3==1.34.0  # For AWS S3
django-storages==1.14.2

# Celery and tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
eventlet==0.33.3  # For Windows

# Payment gateways
stripe==7.5.0
paypalrestsdk==1.13.1

# PDF generation
reportlab==4.1.0
xhtml2pdf==0.2.12

# Excel/CSV export
openpyxl==3.1.2
pandas==2.1.4

# Security
django-axes==6.3.0  # For brute force protection
django-ratelimit==4.1.0

# SEO and sitemaps
django-robots==5.0
django-sitemap==1.0.1

# Debugging and development
django-debug-toolbar==4.2.0
ipython==8.18.1
werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Monitoring and logging
sentry-sdk==1.39.0
prometheus-client==0.19.0